requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
psutil>=5.9.0
schedule>=1.2.0
pytz>=2023.3
python-dotenv>=1.0.0